- Create folder for helper functions. - Outsource training model to dedicated function.
