+++ .gitignore
@@ -1,2 +1,3 @@
 # numpy arrays
 *.npy
+Notes.txt
